Ansible configuration for Nvidia, Opencv and Caffe in a Debian Jessie machine
There are 5 roles. nvidia, opencv and caffe are installed in that order as root
and they install all the dependencies.

The other 2 roles, virtual_opencv and virtual_caffe, are used for install in
a virtual environment both libraries. The defaults options for these tasks are
available for all users. In this files you must declare the path to the virtual
environment and the path for the libraries.